requirements.txt
neat-python==0.92
numpy==1.19.5
pygame==2.0.1
